+++ src/MyLab.WebErrors/HttpTraceIdFact.cs
@@ -1,0 +1,9 @@
+﻿using YamlDotNet.Core.Tokens;
+
+namespace MyLab.WebErrors
+{
+    public static class HttpTraceIdFact
+    {
+        public const string Key = "http-trace-id";
+    }
+}