+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file
 
 Формат лога изменений базируется на [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [1.1.3] - 2020-06-22
+
+### Added
+
+* Log fact `http-trace-id` when unhandled exception cought
+* Update reference to `MyLab.Log.Dsl`
+
 ## [1.0.3] - 2020-05-12
 
 ### Changed